[vscode] Fix unintended JSON overrides
The built-in JSON extension does not respect language overrides,
preventing us from settings defaults we want to apply to every file but
JSON (e.g., editor.formatOnSave and files.insertFinalNewline). With that
disabled, we can now set these (and in particular can get rid of
workarounds like the nu command of `file save-with-newline`). We also
set `"editor.tabSize": 2` while we're at it for JSON.