[nu] Properly use vendor autoload directory
Things put in an vendor autoload directory are *sourced* (and not *used*), so
it was previously being misused. Leveraging this, we can now get rid of relying
on running nu in "login" mode, effectively symlinking the contents of a file
similar to login.nu to a vendor autoload directory. We can also now put
OS-specific modules in OS-specific files and then conditionally symlink those in
bootstrap/nu.sh.